requirements.txt
astroid==2.2.5
attrs==19.1.0
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
codecov==2.0.15
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
Django==2.2.3
django-nose==1.4.6
django-phonenumber-field==3.0.1
django-rest-swagger==2.2.0
djangorestframework==3.10.1
djangorestframework-gis==0.14
drf-yasg==1.16.1
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
nose==1.3.7
openapi-codec==1.3.2
packaging==19.0
phonenumbers==8.10.15
psycopg2-binary==2.8.3
pylint==2.3.1
pyparsing==2.4.1
pyrsistent==0.15.3
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
ruamel.yaml==0.15.100
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
